fix: 'stddef' provide size_t type in config.h
It was observed that if the 'config.h' file was placed in the 'server.c' file as the first include, the compiler could not find the definition of the size_t type. This was not the case when the include occurred after 'stdio.h'. However, the C formatter automatically moves the include of 'config.h' to the top of the 'server.c' includes (alphabetically). To avoid problems in every possible situation, it was decided to include the definition of size_t directly in the 'config.h' file, via 'stddef.h'.