fix: probably the most sneaky bug of my entire C programming experience, so far
By creating the 'rwlock' object via the 'rwlock_create' method, at the time of memory allocation using malloc, the lack of '_t' at the end of the identifier passed as argument to sizeof, indicating the type 'rwlock_t' (valid and defined in the header 'rwlock.h'), led the program to segmentation fault, memory corruption and so on. Unfortunately, the fact that the identifier of the object being created was exactly 'rwlock', led the compiler not to generate errors, as everything was syntactically correct. Thanks valgrind for your help.